Address PR review: use int64 for counts, add parseInt64, add userplaycount to fixtures
Agent-Logs-Url: https://github.com/marcocampos/scrobble-go/sessions/8d40b07d-f6f1-4734-b490-3c9d57e34fd0

+++ examples/album-info/main.go
@@ -92,7 +92,7 @@ func orNone(s string) string {
 	return s
 }
 
-func formatInt(n int) string {
+func formatInt(n int64) string {
 	s := fmt.Sprintf("%d", n)
 	if len(s) <= 3 {
 		return s
